Text files used in final project
These are all of the text files I used. Both the ones used to read
things in and write things are included here. I don't know how to make
sure that you are able to access them all but I think this was probably
as close as I could get to making sure you're able to